+++ lib/enver/boolean_parser.ex
@@ -1,8 +1,4 @@
 defmodule Enver.BooleanParser do
-  @moduledoc """
-  `Enver.BinaryParser` is trying really hard to add value :P
-  """
-
   @type invalid :: Enver.invalid()
   @type opts :: Enver.parse_opts()
   @type val :: Enver.val()
@@ -12,8 +8,14 @@ defmodule Enver.BooleanParser do
   # API #
   #######
 
+  @doc """
+  Converts the given binary to an boolean.
+
+      iex(1)> Enver.BooleanParser.parse("true", %{type: :boolean})
+      {:ok, true}
+  """
   @spec parse(val(), opts()) :: valid() | invalid()
-  def parse(val, _) when is_binary(val) do
+  def parse(val, %{type: :boolean}) when is_binary(val) do
     case String.downcase(val) do
       "false" ->
         {:ok, false}