+++ lib/enver/charlist_parser.ex
@@ -1,5 +1,4 @@
 defmodule Enver.CharlistParser do
-  @type invalid :: Enver.invalid()
   @type opts :: Enver.parse_opts()
   @type val :: Enver.val()
   @type valid :: {:ok, charlist()}
@@ -8,12 +7,14 @@ defmodule Enver.CharlistParser do
   # API #
   #######
 
-  @spec parse(val(), opts()) :: valid() | invalid()
-  def parse("", _) do
-    {:error, "invalid charlist"}
-  end
+  @doc """
+  Converts the given binary to an charlist.
 
-  def parse(val, _) when is_binary(val) do
+      iex(1)> Enver.CharlistParser.parse("potato", %{type: :charlist})
+      {:ok, 'potato'}
+  """
+  @spec parse(val(), opts()) :: valid()
+  def parse(val, %{type: :charlist}) when is_binary(val) do
     {:ok, to_charlist(val)}
   end
 end