+++ test/enver/charlist_parser_test.exs
@@ -1,10 +1,14 @@
 defmodule Enver.CharlistParserTest do
   use ExUnit.Case, async: true
 
-  @parse &Enver.CharlistParser.parse/2
+  @subject Enver.CharlistParser
+
+  def opts(%{} = new_opts \\ %{}) do
+    Map.merge(%{type: :charlist}, new_opts)
+  end
 
   test "parsing returns error for invalid charlist" do
-    opts = %{type: :charlist}
-    assert @parse.("", opts) == {:error, "invalid charlist"}
+    opts = opts()
+    assert @subject.parse("potato", opts) == {:ok, 'potato'}
   end
 end