+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Changelog
 
-# v1.1 (Sat Aug 25 21:55:43 2018 -0700)
+# v0.2.0 (Sat Aug 25 21:55:43 2018 -0700)
+
+- add validations to float parsing
+  - at_least
+  - at_most
+  - greater_than
+  - less_than
+- add validations to integer parsing
+  - greater_than
+  - less_than
+- add type checking to atom parser
+- add type checking to binary parser
+- add type checking to boolean parser
+- add type checking to charlist parser
+- add type checking to float parser
+- add type checking to integer parser
+
+# v0.1.0 (Sat Aug 25 21:55:43 2018 -0700)
 
 - add atom parsing
 - add binary parsing